+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.7] - 2023-08-21
+### Fixed
+- Fixed 24 hours time format (YouTube).
+- Fixed duplicated setting button (YouTube).
+
 ## [1.3.6] - 2023-08-18
 ### Fixed
 - Fixed option menu (YouTube).